+++ test/index.test.js
@@ -598,6 +598,7 @@ describe("🇺🇦 Ukrainian", () => {
   it("ukrainianCyrillicToLatynka", () => {
     expect(ukrainianCyrillicToLatynka("Міль"), "Milj");
     expect(ukrainianCyrillicToLatynka("Є İНШИЙ"), "JE INŠYJ");
+    expect(ukrainianCyrillicToLatynka("Я БУВ ЗАЙНЯТИЙ ОСТАННІМ ЧАСОМ"), "JA BUV ZAJNJATYJ OSTANNIM ČASOM");
   });
   it("ukrainianLatynkaToCyrillic", () => {
     expect(ukrainianLatynkaToCyrillic("Milj"), "Міль");